Initial version of cmake build support
This follows pretty closely the contents of the old
Makefile.config.template file. All options can be specified in the
command-line, so there should be no need to modify anything to compile
different flavours of stf.

The following points are still to be addressed:

 * The stf binary doesn't have a different name depending on the flavour
 * The treefrog software has been left out
 * Automatic ADIOS discovery has been left out
 * Documentation generation is not included yet

Other than that, the software seems to be compiling fine and picking up
all dependencies (GSL, HDF5, OpenMP and MPI). I've tested it at least
against g++ and clang++, with/without OpenMP and MPI, and the software
seems to be compiling fine in all cases.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>