First attempt to have nbody-specific flags
All nbody-specific macro definitions (with the big exceptio of USEMPI)
are now defined in the NBodyLib CMakeLists.txt file, and are now called
NBODY_*. They are also exported to the parent scope so programs that are
include nbodylib as part of their compilation (like VR and TF) can see
these definitions and use them as well. This is of course necessary,
because otherwise the NBodyLib header files would not correspond with
what was compiled in the first place.

A similar approach has been taken with VELOCIraptor, which also defines
a number of macros that are of interest to TreeFrog, if one is building
TreeFrog together with VELOCIraptor.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>