Separating main target into STATIC and INTERFACE
Making this separation allows us to keep generating the static
libnbodylib.a library with all the objects, but also to expose these
objects to be contained in other static libraries. This is needed by
velociraptor, which needs to generate a big, fat libvelociraptor.a
static library so in turn Swift can link against it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>