Add new gsl_invoke helper function
This function is meant to be used by gsl invocations returning an
integer status. Whenever this status is other than GSL_SUCCESS we throw
a gsl_error exception, which will safely propagate through the C++
stack.

This of course needs users to set the "off" error handler before any
other GSL routine is called, but that's up to the users to do.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>